+++ src/main/java/com/mrfti/erp/domain/dtos/ProdutoDTO.java
@@ -1,7 +1,6 @@
 package com.mrfti.erp.domain.dtos;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 import com.mrfti.erp.domain.CategoriaProduto;
 import com.mrfti.erp.domain.Produto;
@@ -13,7 +12,7 @@ public class ProdutoDTO implements Serializable {
 
 	protected Integer id;
 	protected String descricao;
-	protected BigDecimal preco;
+	protected Double preco;
 	
 	protected CategoriaProduto categoriaProduto;
 	
@@ -47,11 +46,11 @@ public class ProdutoDTO implements Serializable {
 		this.descricao = descricao;
 	}
 
-	public BigDecimal getPreco() {
+	public Double getPreco() {
 		return preco;
 	}
 
-	public void setPreco(BigDecimal preco) {
+	public void setPreco(Double preco) {
 		this.preco = preco;
 	}
